correccion double por floar

+++ app/src/main/java/es/jrodg85/realestate/OfertaDeAlquiler.java
@@ -22,7 +22,7 @@ public class OfertaDeAlquiler extends Oferta implements OfertaDeAlquilerInterfaz
   // setters
 
   @Override
-  public void setPrecioAlquilerMensual(float precioAlquilerMensual) {
+  public void setPrecioAlquilerMensual(double precioAlquilerMensual) {
     this.precioAlquilerMensual = precioAlquilerMensual;
   }
